projects/django-ecommerce/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
cryptography==45.0.7
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.6
django-allauth==0.57.0
django-cors-headers==4.3.1
django-filter==23.5
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==37.6.0
flake8==6.1.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==7.8.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
